requirements.txt
pyyaml~=6.0
pypresence~=4.2.1
praw~=7.3.0
packaging~=21.0
requests~=2.27.1
